pyproject.toml
[project]
name = "amplifier-bundle-digital-twin-universe"
version = "99.0.0"
description = "On-demand ephemeral Digital Twin Universe environments from declarative profiles"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Microsoft MADE:Explorations Team" }]
keywords = ["amplifier", "digital-twin", "incus", "ephemeral", "simulation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.3,<9.0",
    "pyyaml>=6.0,<7.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff",
    "pyright",
]

[project.scripts]
amplifier-digital-twin = "amplifier_bundle_digital_twin_universe.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/amplifier_bundle_digital_twin_universe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["tests"]
markers = [
    "integration: tests that require Incus (deselected by default, use --run-integration)",
    "e2e: full end-to-end tests (requires Docker, Incus, API keys, use --run-e2e)",
]